requirements.txt
discord.py==2.5.2
spotipy==2.25.1
wavelink==3.4.1
colorlog==6.9.0
